Push the v5.4 tag of the image to quay
Push a new tag v5.4, so that podman-client in main branch can start using it as the default
tag. Continue pushing tag v5.3 too because the GA version of 5.3 hasn't been released yet.

When Podman v5.3 is in Fedora we should stop pushing tag v5.3 of the image and
start pushing v5.5 along with v5.4.

Signed-off-by: Mario Loriedo <mario.loriedo@gmail.com>